feat: remove user from board team when tab closes
