+++ src/app/app.component.ts
@@ -1,4 +1,5 @@
-import { Component } from "@angular/core";
+import { Component, HostListener } from "@angular/core";
+import { BoardService } from "./services/board.service";
 
 @Component({
   selector: "app-root",
@@ -6,4 +7,19 @@ import { Component } from "@angular/core";
 })
 export class AppComponent {
   title = "retrocket";
+
+  /* Remove user from team of users accessing the board */
+  @HostListener("window:beforeunload", ["$event"])
+  public beforeunloadHandler() {
+    const currentUser = JSON.parse(localStorage.getItem("user"));
+    if (currentUser)
+      this.boardService.updateBoardTeam(currentUser.lastJoined, {
+        type: "remove",
+        member: currentUser.displayName,
+      });
+
+    localStorage.removeItem("user");
+  }
+
+  constructor(private boardService: BoardService) {}
 }
