+++ src/app/components/auth/auth.component.html
@@ -1,4 +1,4 @@
-<div class="header bg-gradient-warning py-5 py-lg-6">
+<div class="header bg-gradient-primary py-5 py-lg-6">
   <div class="container">
     <div class="header-body text-center mt--4 mb-7">
       <div class="row justify-content-center">
@@ -21,18 +21,18 @@
 <div class="container mt--8 pb-5">
   <div class="row justify-content-center">
     <div class="col-lg-5 col-md-7">
-      <div class="card bg-secondary shadow" style="padding: 15px">
-        <div *ngIf="loggedIn">
+      <div class="card bg-secondary shadow" style="padding: 50px 5px">
+        <div *ngIf="isAuthor">
           <div class="text-center">
             <div class="text-center my-3">
               <h3>Done for the day?</h3>
             </div>
-            <button (click)="onClickLogout()" class="btn btn-warning my-4">
+            <button (click)="onClickLogout()" class="btn auth my-4">
               <i class="fas fa-sign-out-alt"></i>
               Sign out
             </button>
-            <a class="greet" [routerLink]="['/boards']">
-              <button class="btn btn-warning my-4">
+            <a class="greet" [routerLink]="['/dashboard']">
+              <button class="btn auth my-4">
                 <i class="fas fa-laugh"></i>
                 Back to Dashboard
               </button>
