+++ src/app/components/board/board.component.html
@@ -1,0 +1,39 @@
+<app-menubar
+  [color]="color$ | async"
+  [colors]="colors$ | async"
+  [team]="(board$ | async)?.team"
+  [boardId]="boardId"
+  [boardName]="(board$ | async)?.name"
+  (boardUpdate)="updateBoard($event)"
+></app-menubar>
+<div>
+  <div class="header pb-8 pt-5" [ngClass]="(color$ | async)?.name">
+    <div class="container-fluid">
+      <div class="header-body">
+        <div class="row">
+          <div
+            class="col-xl-3 col-lg-6"
+            *ngFor="let column of columns$ | async"
+          >
+            <app-column-header
+              [column]="column"
+              [count]="cards[column.order]?.length"
+            ></app-column-header>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+  <div class="container-fluid mt--8 mb-5 px-5">
+    <div class="row mt-5">
+      <div class="col-xl-3 col-lg-6" *ngFor="let column of columns$ | async">
+        <app-column
+          [column]="column"
+          [cards]="cards[column.order]"
+          (update)="updateColumn($event)"
+          (cardEvent)="handleCardEvent($event)"
+        ></app-column>
+      </div>
+    </div>
+  </div>
+</div>
