--- src/app/components/dashboard/dashboard-list.component.html
@@ -1,62 +1,0 @@
-<div class="header bg-gradient-warning py-5 py-lg-6">
-  <div class="container">
-    <div class="header-body text-center mt--4 mb-7">
-      <div class="justify-content-center">
-        <div class="col-lg-5 col-md-6">
-          <a class="navbar-brand">
-            <img class="logo" src="assets/img/brand/retrocket-white.png" />
-          </a>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-<div *ngIf="boards" class="container-fluid mt--7">
-  <div class="row">
-    <div class="col rt-table">
-      <div class="card shadow">
-        <div class="card-header">
-          <div class="row">
-            <div class="col">
-              <h2 class="mb-0">Past Retrospective Sessions</h2>
-            </div>
-            <div class="col">
-              <button
-                type="button"
-                class="btn btn-lg btn-outline-default float-right"
-                (click)="onClickSignout()"
-              >
-                <i class="fas fa-plus-circle"></i>Sign out
-              </button>
-              <button
-                type="button"
-                class="btn btn-lg btn-outline-default float-right"
-                (click)="openDialog()"
-              >
-                <i class="fas fa-plus-circle"></i>New Session
-              </button>
-            </div>
-          </div>
-        </div>
-        <div class="table-responsive">
-          <table class="table align-items-center table-flush">
-            <thead class="thead-light">
-              <tr>
-                <th style="width: 30%" scope="col">Session</th>
-                <th style="width: 20%" scope="col">Created</th>
-                <th style="width: 40%" scope="col">Memo</th>
-                <th style="width: 10%" scope="col">Actions</th>
-              </tr>
-            </thead>
-            <tbody>
-              <ng-container *ngFor="let board of boards"
-                ><tr app-dashboard-list-item [board]="board"></tr>
-              </ng-container>
-            </tbody>
-          </table>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-<router-outlet></router-outlet>
