+++ src/app/components/dashboard/dashboard.component.scss
@@ -1,0 +1,31 @@
+div.rt-table {
+  margin-bottom: 20px;
+  max-width: 90%;
+  margin: auto;
+}
+
+.header-body {
+  justify-content: center;
+  display: flex;
+}
+
+h1 {
+  color: white;
+  font-size: xxx-large;
+}
+
+.btn {
+  margin-right: 10px;
+  margin-left: 10px;
+}
+
+table {
+  table-layout: fixed;
+  width: 100%;
+  white-space: nowrap;
+}
+
+.table-responsive {
+  overflow-y: visible;
+  overflow-x: visible;
+}
