+++ src/app/components/menubar/menubar.component.html
@@ -4,13 +4,13 @@
   id="sidenav-main"
 >
   <div class="container-fluid bg-gradient-neutral">
-    <a class="navbar-brand" [routerLink]="['/boards']">
+    <a class="navbar-brand" [routerLink]="['/dashboard']">
       <img src="./assets/img/brand/retrocket.png" class="logo" />
     </a>
     <div class="navbar-collapse row">
-      <ul class="navbar-nav">
+      <ul class="navbar-nav text-uppercase">
         <li class="nav-item">
-          <a [routerLink]="['/boards']" class="nav-link">
+          <a [routerLink]="['/dashboard']" class="nav-link">
             <i class="fas fa-columns"></i>
             Dashboard
           </a>
